Created tests for chart5y component mount and data load
