+++ client/components/app.jsx
@@ -115,7 +115,7 @@ class App extends React.Component {
     var change = (this.state.selectedPrice - this.state.refStartPrice).toFixed(2);
     var changePercent = ((this.state.selectedPrice - this.state.refStartPrice) * 100 / this.state.refStartPrice).toFixed(2);
     return (
-      <div id='test'>
+      <div>
         <CompanyName>{this.state.companyName}</CompanyName>
         <div><Odometer value={this.state.selectedPrice} format='(,ddd).dd' duration={300}></Odometer></div>
         <div><Change>{change > 0 ? '+$' : '-$'} {Math.abs(change)} {'(' + changePercent + '%) '}</Change><ChangeCaption>{this.state.changeCaption}</ChangeCaption></div>
