+++ package.json
@@ -9,7 +9,7 @@
     "node": ">=6.13.0"
   },
   "jest": {
-    "setupTestFrameworkScriptFile": "<rootDir>/setupTests.js"
+    "setupTestFrameworkScriptFile": "<rootDir>/tests/setupTests.js"
   },
   "scripts": {
     "start": "nodemon server/index.js",
@@ -22,7 +22,6 @@
     "express": "^4.16.4",
     "faker": "^4.1.0",
     "highcharts": "^7.0.1",
-    "jsdom": "^13.1.0",
     "mongoose": "^5.4.0",
     "nodemon": "^1.18.9",
     "react": "^16.7.0",
