+++ tests/setupTests.js
@@ -1,0 +1,42 @@
+// ===========================================
+// CONFIGURE ENZYME FOR REACT VERSION
+// ===========================================
+var enzyme = require('enzyme');
+var Adapter = require('enzyme-adapter-react-16');
+
+enzyme.configure({ adapter: new Adapter() });
+
+
+// ===========================================
+// STUB FUNCTIONS AND MOCK DATA
+// ===========================================
+
+var mock5yPrices = [
+  {"datetime":"2018-12-27T02:00:24.342Z","price":1461.99},{"datetime":"2018-12-20T02:00:24.342Z","price":1459.76},{"datetime":"2018-12-13T02:00:24.342Z","price":1488.64},{"datetime":"2018-12-06T02:00:24.342Z","price":1511.34},{"datetime":"2018-11-29T02:00:24.342Z","price":1531.95},{"datetime":"2018-11-22T02:00:24.342Z","price":1501.51},{"datetime":"2018-11-15T02:00:24.342Z","price":1518.23},{"datetime":"2018-11-08T02:00:24.342Z","price":1531.07},{"datetime":"2018-11-01T01:00:24.342Z","price":1550.15},{"datetime":"2018-10-25T01:00:24.342Z","price":1555.1},{"datetime":"2018-10-18T01:00:24.342Z","price":1534.25},{"datetime":"2018-10-11T01:00:24.342Z","price":1515.85},{"datetime":"2018-10-04T01:00:24.342Z","price":1501.64},{"datetime":"2018-09-27T01:00:24.342Z","price":1529.59},{"datetime":"2018-09-20T01:00:24.342Z","price":1528.06},{"datetime":"2018-09-13T01:00:24.342Z","price":1540.07},{"datetime":"2018-09-06T01:00:24.342Z","price":1561},{"datetime":"2018-08-30T01:00:24.342Z","price":1551.65},{"datetime":"2018-08-23T01:00:24.342Z","price":1548.32},{"datetime":"2018-08-16T01:00:24.342Z","price":1553.36},{"datetime":"2018-08-09T01:00:24.342Z","price":1552.57},{"datetime":"2018-08-02T01:00:24.342Z","price":1541.38},{"datetime":"2018-07-26T01:00:24.342Z","price":1544.17},{"datetime":"2018-07-19T01:00:24.342Z","price":1548.13},{"datetime":"2018-07-12T01:00:24.342Z","price":1566.72},{"datetime":"2018-07-05T01:00:24.342Z","price":1554.72},{"datetime":"2018-06-28T01:00:24.342Z","price":1543.49},{"datetime":"2018-06-21T01:00:24.342Z","price":1547.19},{"datetime":"2018-06-14T01:00:24.342Z","price":1547.86},{"datetime":"2018-06-07T01:00:24.342Z","price":1559.09},{"datetime":"2018-05-31T01:00:24.342Z","price":1543.08},{"datetime":"2018-05-24T01:00:24.342Z","price":1563.85},{"datetime":"2018-05-17T01:00:24.342Z","price":1543.25},{"datetime":"2018-05-10T01:00:24.342Z","price":1558.72},{"datetime":"2018-05-03T01:00:24.342Z","price":1580.51},{"datetime":"2018-04-26T01:00:24.342Z","price":1556.99},{"datetime":"2018-04-19T01:00:24.342Z","price":1540.21},{"datetime":"2018-04-12T01:00:24.342Z","price":1560.67},{"datetime":"2018-04-05T01:00:24.342Z","price":1544.97},{"datetime":"2018-03-29T01:00:24.342Z","price":1534.17},{"datetime":"2018-03-22T01:00:24.342Z","price":1528.03},{"datetime":"2018-03-15T01:00:24.342Z","price":1549.61},{"datetime":"2018-03-08T02:00:24.342Z","price":1578.07},{"datetime":"2018-03-01T02:00:24.342Z","price":1589.07},{"datetime":"2018-02-22T02:00:24.342Z","price":1596.16},{"datetime":"2018-02-15T02:00:24.342Z","price":1570.75},{"datetime":"2018-02-08T02:00:24.342Z","price":1547.18},{"datetime":"2018-02-01T02:00:24.342Z","price":1564.98},{"datetime":"2018-01-25T02:00:24.342Z","price":1552.33},{"datetime":"2018-01-18T02:00:24.342Z","price":1582.9},{"datetime":"2018-01-11T02:00:24.342Z","price":1572.56},{"datetime":"2018-01-04T02:00:24.342Z","price":1584.79},{"datetime":"2017-12-28T02:00:24.342Z","price":1613.58},{"datetime":"2017-12-21T02:00:24.342Z","price":1627.64},{"datetime":"2017-12-14T02:00:24.342Z","price":1608.63},{"datetime":"2017-12-07T02:00:24.342Z","price":1610.91},{"datetime":"2017-11-30T02:00:24.342Z","price":1590.71},{"datetime":"2017-11-23T02:00:24.342Z","price":1612.58},{"datetime":"2017-11-16T02:00:24.342Z","price":1610.81},{"datetime":"2017-11-09T02:00:24.342Z","price":1602.56},{"datetime":"2017-11-02T01:00:24.342Z","price":1602.16},{"datetime":"2017-10-26T01:00:24.342Z","price":1619.49},{"datetime":"2017-10-19T01:00:24.342Z","price":1650.19},{"datetime":"2017-10-12T01:00:24.342Z","price":1658.77},{"datetime":"2017-10-05T01:00:24.342Z","price":1653.16},{"datetime":"2017-09-28T01:00:24.342Z","price":1622.87},{"datetime":"2017-09-21T01:00:24.342Z","price":1628.66},{"datetime":"2017-09-14T01:00:24.342Z","price":1609.74},{"datetime":"2017-09-07T01:00:24.342Z","price":1608.77},{"datetime":"2017-08-31T01:00:24.342Z","price":1584.26},{"datetime":"2017-08-24T01:00:24.342Z","price":1612.4},{"datetime":"2017-08-17T01:00:24.342Z","price":1582.32},{"datetime":"2017-08-10T01:00:24.342Z","price":1557.43},{"datetime":"2017-08-03T01:00:24.342Z","price":1539.09},{"datetime":"2017-07-27T01:00:24.342Z","price":1537.7},{"datetime":"2017-07-20T01:00:24.342Z","price":1565.85},{"datetime":"2017-07-13T01:00:24.342Z","price":1550.43},{"datetime":"2017-07-06T01:00:24.342Z","price":1535.4},{"datetime":"2017-06-29T01:00:24.342Z","price":1552.82},{"datetime":"2017-06-22T01:00:24.342Z","price":1560.62},{"datetime":"2017-06-15T01:00:24.342Z","price":1546.13},{"datetime":"2017-06-08T01:00:24.342Z","price":1571.28},{"datetime":"2017-06-01T01:00:24.342Z","price":1563.34},{"datetime":"2017-05-25T01:00:24.342Z","price":1552.33},{"datetime":"2017-05-18T01:00:24.342Z","price":1536.13},{"datetime":"2017-05-11T01:00:24.342Z","price":1566.39},{"datetime":"2017-05-04T01:00:24.342Z","price":1567.14},{"datetime":"2017-04-27T01:00:24.342Z","price":1590.98},{"datetime":"2017-04-20T01:00:24.342Z","price":1603.05},{"datetime":"2017-04-13T01:00:24.342Z","price":1574.8},{"datetime":"2017-04-06T01:00:24.342Z","price":1591.65},{"datetime":"2017-03-30T01:00:24.342Z","price":1571.13},{"datetime":"2017-03-23T01:00:24.342Z","price":1550.15},{"datetime":"2017-03-16T01:00:24.342Z","price":1524.01},{"datetime":"2017-03-09T02:00:24.342Z","price":1503.12},{"datetime":"2017-03-02T02:00:24.342Z","price":1496.66},{"datetime":"2017-02-23T02:00:24.342Z","price":1518.07},{"datetime":"2017-02-16T02:00:24.342Z","price":1539.32},{"datetime":"2017-02-09T02:00:24.342Z","price":1545.66},{"datetime":"2017-02-02T02:00:24.342Z","price":1551.73},{"datetime":"2017-01-26T02:00:24.342Z","price":1532.28},{"datetime":"2017-01-19T02:00:24.342Z","price":1518.23},{"datetime":"2017-01-12T02:00:24.342Z","price":1494.98},{"datetime":"2017-01-05T02:00:24.342Z","price":1485.5},{"datetime":"2016-12-29T02:00:24.342Z","price":1462.18},{"datetime":"2016-12-22T02:00:24.342Z","price":1469.37},{"datetime":"2016-12-15T02:00:24.342Z","price":1494.15},{"datetime":"2016-12-08T02:00:24.342Z","price":1482.58},{"datetime":"2016-12-01T02:00:24.342Z","price":1476.6},{"datetime":"2016-11-24T02:00:24.342Z","price":1504.16},{"datetime":"2016-11-17T02:00:24.342Z","price":1489.96},{"datetime":"2016-11-10T02:00:24.342Z","price":1462.48},{"datetime":"2016-11-03T01:00:24.342Z","price":1439.68},{"datetime":"2016-10-27T01:00:24.342Z","price":1442.72},{"datetime":"2016-10-20T01:00:24.342Z","price":1467.78},{"datetime":"2016-10-13T01:00:24.342Z","price":1453.09},{"datetime":"2016-10-06T01:00:24.342Z","price":1429.14},{"datetime":"2016-09-29T01:00:24.342Z","price":1434.72},{"datetime":"2016-09-22T01:00:24.342Z","price":1461.79},{"datetime":"2016-09-15T01:00:24.342Z","price":1436.43},{"datetime":"2016-09-08T01:00:24.342Z","price":1458.99},{"datetime":"2016-09-01T01:00:24.342Z","price":1455.66},{"datetime":"2016-08-25T01:00:24.342Z","price":1444.66},{"datetime":"2016-08-18T01:00:24.342Z","price":1439.03},{"datetime":"2016-08-11T01:00:24.342Z","price":1453.22},{"datetime":"2016-08-04T01:00:24.342Z","price":1456.54},{"datetime":"2016-07-28T01:00:24.342Z","price":1444.15},{"datetime":"2016-07-21T01:00:24.342Z","price":1440.89},{"datetime":"2016-07-14T01:00:24.342Z","price":1459.08},{"datetime":"2016-07-07T01:00:24.342Z","price":1436.63},{"datetime":"2016-06-30T01:00:24.342Z","price":1421.78},{"datetime":"2016-06-23T01:00:24.342Z","price":1442.29},{"datetime":"2016-06-16T01:00:24.342Z","price":1449.92},{"datetime":"2016-06-09T01:00:24.342Z","price":1465.75},{"datetime":"2016-06-02T01:00:24.342Z","price":1446.93},{"datetime":"2016-05-26T01:00:24.342Z","price":1455.9},{"datetime":"2016-05-19T01:00:24.342Z","price":1459.05},{"datetime":"2016-05-12T01:00:24.342Z","price":1479.05},{"datetime":"2016-05-05T01:00:24.342Z","price":1475.05},{"datetime":"2016-04-28T01:00:24.342Z","price":1478.7},{"datetime":"2016-04-21T01:00:24.342Z","price":1461.77},{"datetime":"2016-04-14T01:00:24.342Z","price":1449.35},{"datetime":"2016-04-07T01:00:24.342Z","price":1467.93},{"datetime":"2016-03-31T01:00:24.342Z","price":1496.16},{"datetime":"2016-03-24T01:00:24.342Z","price":1468.59},{"datetime":"2016-03-17T01:00:24.342Z","price":1462.71},{"datetime":"2016-03-10T02:00:24.342Z","price":1465.49},{"datetime":"2016-03-03T02:00:24.342Z","price":1436.25},{"datetime":"2016-02-25T02:00:24.342Z","price":1431.46},{"datetime":"2016-02-18T02:00:24.342Z","price":1422.94},{"datetime":"2016-02-11T02:00:24.342Z","price":1417.33},{"datetime":"2016-02-04T02:00:24.342Z","price":1441.33},{"datetime":"2016-01-28T02:00:24.342Z","price":1445.99},{"datetime":"2016-01-21T02:00:24.342Z","price":1418.84},{"datetime":"2016-01-14T02:00:24.342Z","price":1392.64},{"datetime":"2016-01-07T02:00:24.342Z","price":1376.56},{"datetime":"2015-12-31T02:00:24.342Z","price":1360.95},{"datetime":"2015-12-24T02:00:24.342Z","price":1353.45},{"datetime":"2015-12-17T02:00:24.342Z","price":1360.73},{"datetime":"2015-12-10T02:00:24.342Z","price":1368.84},{"datetime":"2015-12-03T02:00:24.342Z","price":1374.26},{"datetime":"2015-11-26T02:00:24.342Z","price":1359.53},{"datetime":"2015-11-19T02:00:24.342Z","price":1378.8},{"datetime":"2015-11-12T02:00:24.342Z","price":1354.75},{"datetime":"2015-11-05T02:00:24.342Z","price":1337.44},{"datetime":"2015-10-29T01:00:24.342Z","price":1312.9},{"datetime":"2015-10-22T01:00:24.342Z","price":1286.84},{"datetime":"2015-10-15T01:00:24.342Z","price":1289.69},{"datetime":"2015-10-08T01:00:24.342Z","price":1295.29},{"datetime":"2015-10-01T01:00:24.342Z","price":1305.71},{"datetime":"2015-09-24T01:00:24.342Z","price":1286.33},{"datetime":"2015-09-17T01:00:24.342Z","price":1286.23},{"datetime":"2015-09-10T01:00:24.342Z","price":1299.3},{"datetime":"2015-09-03T01:00:24.342Z","price":1280.43},{"datetime":"2015-08-27T01:00:24.342Z","price":1300.8},{"datetime":"2015-08-20T01:00:24.342Z","price":1307.69},{"datetime":"2015-08-13T01:00:24.342Z","price":1288.89},{"datetime":"2015-08-06T01:00:24.342Z","price":1293.19},{"datetime":"2015-07-30T01:00:24.342Z","price":1293.13},{"datetime":"2015-07-23T01:00:24.342Z","price":1318.16},{"datetime":"2015-07-16T01:00:24.342Z","price":1310.45},{"datetime":"2015-07-09T01:00:24.342Z","price":1295.86},{"datetime":"2015-07-02T01:00:24.342Z","price":1292.23},{"datetime":"2015-06-25T01:00:24.342Z","price":1276.81},{"datetime":"2015-06-18T01:00:24.342Z","price":1266.94},{"datetime":"2015-06-11T01:00:24.342Z","price":1252.86},{"datetime":"2015-06-04T01:00:24.342Z","price":1228.84},{"datetime":"2015-05-28T01:00:24.342Z","price":1220.64},{"datetime":"2015-05-21T01:00:24.342Z","price":1214.25},{"datetime":"2015-05-14T01:00:24.342Z","price":1234.39},{"datetime":"2015-05-07T01:00:24.342Z","price":1257.45},{"datetime":"2015-04-30T01:00:24.342Z","price":1256.3},{"datetime":"2015-04-23T01:00:24.342Z","price":1254.49},{"datetime":"2015-04-16T01:00:24.342Z","price":1264.15},{"datetime":"2015-04-09T01:00:24.342Z","price":1285.31},{"datetime":"2015-04-02T01:00:24.342Z","price":1281.97},{"datetime":"2015-03-26T01:00:24.342Z","price":1286.32},{"datetime":"2015-03-19T01:00:24.342Z","price":1308.28},{"datetime":"2015-03-12T01:00:24.342Z","price":1315.91},{"datetime":"2015-03-05T02:00:24.342Z","price":1337.19},{"datetime":"2015-02-26T02:00:24.342Z","price":1323.4},{"datetime":"2015-02-19T02:00:24.342Z","price":1320.54},{"datetime":"2015-02-12T02:00:24.342Z","price":1294.92},{"datetime":"2015-02-05T02:00:24.342Z","price":1278.48},{"datetime":"2015-01-29T02:00:24.342Z","price":1263.22},{"datetime":"2015-01-22T02:00:24.342Z","price":1274.59},{"datetime":"2015-01-15T02:00:24.342Z","price":1266.4},{"datetime":"2015-01-08T02:00:24.342Z","price":1290.56},{"datetime":"2015-01-01T02:00:24.342Z","price":1278.24},{"datetime":"2014-12-25T02:00:24.342Z","price":1300.07},{"datetime":"2014-12-18T02:00:24.342Z","price":1310.14},{"datetime":"2014-12-11T02:00:24.342Z","price":1334.86},{"datetime":"2014-12-04T02:00:24.342Z","price":1308.66},{"datetime":"2014-11-27T02:00:24.342Z","price":1283.94},{"datetime":"2014-11-20T02:00:24.342Z","price":1259.71},{"datetime":"2014-11-13T02:00:24.342Z","price":1265.34},{"datetime":"2014-11-06T02:00:24.342Z","price":1269.18},{"datetime":"2014-10-30T01:00:24.342Z","price":1293.66},{"datetime":"2014-10-23T01:00:24.342Z","price":1290.5},{"datetime":"2014-10-16T01:00:24.342Z","price":1267.64},{"datetime":"2014-10-09T01:00:24.342Z","price":1254.85},{"datetime":"2014-10-02T01:00:24.342Z","price":1257.11},{"datetime":"2014-09-25T01:00:24.342Z","price":1280.45},{"datetime":"2014-09-18T01:00:24.342Z","price":1270.89},{"datetime":"2014-09-11T01:00:24.342Z","price":1279.02},{"datetime":"2014-09-04T01:00:24.342Z","price":1294.54},{"datetime":"2014-08-28T01:00:24.342Z","price":1274.93},{"datetime":"2014-08-21T01:00:24.342Z","price":1276.34},{"datetime":"2014-08-14T01:00:24.342Z","price":1283.69},{"datetime":"2014-08-07T01:00:24.342Z","price":1263.73},{"datetime":"2014-07-31T01:00:24.342Z","price":1254.04},{"datetime":"2014-07-24T01:00:24.342Z","price":1242.76},{"datetime":"2014-07-17T01:00:24.342Z","price":1252.9},{"datetime":"2014-07-10T01:00:24.342Z","price":1251.02},{"datetime":"2014-07-03T01:00:24.342Z","price":1246.78},{"datetime":"2014-06-26T01:00:24.342Z","price":1234.79},{"datetime":"2014-06-19T01:00:24.342Z","price":1215.25},{"datetime":"2014-06-12T01:00:24.342Z","price":1215.35},{"datetime":"2014-06-05T01:00:24.342Z","price":1205.08},{"datetime":"2014-05-29T01:00:24.342Z","price":1193.14},{"datetime":"2014-05-22T01:00:24.342Z","price":1214.04},{"datetime":"2014-05-15T01:00:24.342Z","price":1233},{"datetime":"2014-05-08T01:00:24.342Z","price":1241.17},{"datetime":"2014-05-01T01:00:24.342Z","price":1237.72},{"datetime":"2014-04-24T01:00:24.342Z","price":1243.24},{"datetime":"2014-04-17T01:00:24.342Z","price":1267.19},{"datetime":"2014-04-10T01:00:24.342Z","price":1253.39},{"datetime":"2014-04-03T01:00:24.342Z","price":1275.92},{"datetime":"2014-03-27T01:00:24.342Z","price":1272.01},{"datetime":"2014-03-20T01:00:24.342Z","price":1286.64},{"datetime":"2014-03-13T01:00:24.342Z","price":1269},{"datetime":"2014-03-06T02:00:24.342Z","price":1277.48},{"datetime":"2014-02-27T02:00:24.342Z","price":1294},{"datetime":"2014-02-20T02:00:24.342Z","price":1295.49},{"datetime":"2014-02-13T02:00:24.342Z","price":1295.47},{"datetime":"2014-02-06T02:00:24.342Z","price":1304.72},{"datetime":"2014-01-30T02:00:24.342Z","price":1327.29},{"datetime":"2014-01-23T02:00:24.342Z","price":1310.29},{"datetime":"2014-01-16T02:00:24.342Z","price":1323.86},{"datetime":"2014-01-09T02:00:24.342Z","price":1303.74}
+];
+
+var mockRequestData = (dataType, callback) => {
+  
+  if (dataType === 'last5yPrices') {
+    window.setTimeout(() => {callback(mock5yPrices);}, 0);
+  }
+};
+
+export {mockRequestData, mock5yPrices};
+
+
+// REFLECTION - ENZYME SETUP 
+// Jest setup was straight forward because it didn't test JSX
+// Enzyme setup was more difficult because React code contains JSX but Jest does not understand JSX by default
+// (0) What special things can Enzyme do? Consider if you want to render a component with certain props and states and ensure certain assertions are true. How would you do that?
+// (1) install Jest if not yet installed: https://jestjs.io/docs/en/getting-started
+// (2) install enzyme: https://airbnb.io/enzyme/docs/installation/
+// (3) Enzyme does not know by default what version of React it should test for. You will need to tell it.
+  // Create a setupTests.js file which configures Enzyme with appropriate version of React: https://airbnb.io/enzyme/docs/installation/
+  // Then, configure jest to run the setup file prior to every test: https://airbnb.io/enzyme/docs/guides/jest.html
+    // Optional reading: https://jestjs.io/docs/en/configuration#setuptestframeworkscriptfile-string
+// (4) Jest does not understand JSX by default. It needs to be setup to transpile JSX through babel. Check your package.json file to determine what version of babel you are using and follow the instructions for the relevant version: https://jestjs.io/docs/en/getting-started#using-babel
+  // Make sure the 'presets' property in your .babelrc file references the correct preset files. For me it was '@babel/preset-env' and 'babel-preset-react', not 'env' and 'react'.
+// (5) Write your first tests!
+  // Sample Jest tests: https://jestjs.io/docs/en/using-matchers
+  // Sample Enzyme tests: https://github.com/vjwilson/enzyme-example-jest/blob/master/src/__tests__/Foo-test.js
